+++ e2e-test/src/message_hub.rs
@@ -16,16 +16,20 @@
 //
 // SPDX-License-Identifier: Apache-2.0
 
-use std::{net::Ipv6Addr, path::Path, time::Duration};
+use std::{net::Ipv6Addr, path::Path, sync::Arc, time::Duration};
 
 use astarte_device_sdk::{
     builder::DeviceBuilder, prelude::*, store::SqliteStore, transport::mqtt::MqttConfig,
 };
-use astarte_message_hub::{init_pub_sub, AstarteMessageHub};
+use astarte_message_hub::{astarte::handler::init_pub_sub, AstarteMessageHub};
 use astarte_message_hub_proto::message_hub_server::MessageHubServer;
 use eyre::{Context, OptionExt};
-use tokio::task::{AbortHandle, JoinSet};
-use tower::ServiceBuilder;
+use futures::{future::BoxFuture, FutureExt};
+use tokio::{
+    sync::Barrier,
+    task::{AbortHandle, JoinSet},
+};
+use tower::{Layer, Service, ServiceBuilder};
 use tower_http::trace::TraceLayer;
 use tracing::instrument;
 
@@ -48,6 +52,7 @@ impl MsgHub {
 #[instrument(skip_all)]
 pub async fn init_message_hub(
     path: &Path,
+    barrier: Arc<Barrier>,
     tasks: &mut JoinSet<eyre::Result<()>>,
 ) -> eyre::Result<MsgHub> {
     let realm = read_env("E2E_REALM")?;
@@ -55,7 +60,8 @@ pub async fn init_message_hub(
     let credentials_secret = read_env("E2E_CREDENTIAL_SECRET")?;
     let pairing_url = read_env("E2E_PAIRING_URL")?;
 
-    let mut mqtt_config = MqttConfig::new(realm, device_id, credentials_secret, pairing_url);
+    let mut mqtt_config =
+        MqttConfig::with_credential_secret(realm, device_id, credentials_secret, pairing_url);
 
     if read_env("E2E_IGNORE_SSL").is_ok() {
         mqtt_config.ignore_ssl_errors();
@@ -64,26 +70,28 @@ pub async fn init_message_hub(
     let path = path.to_str().ok_or_eyre("invalid_path")?;
 
     let uri = format!("sqlite://{path}/store.db");
-    let store = SqliteStore::new(&uri).await?;
+    let store = SqliteStore::from_uri(&uri).await?;
 
-    let (mut device, rx_events) = DeviceBuilder::new()
+    let (client, mut connection) = DeviceBuilder::new()
         .store(store)
         .connect(mqtt_config)
         .await?
         .build();
 
-    let (publisher, mut subscriber) = init_pub_sub(device.clone(), rx_events);
+    let (publisher, mut subscriber) = init_pub_sub(client);
 
     let message_hub = AstarteMessageHub::new(publisher);
 
-    let server = tasks.spawn(async {
+    let server = tasks.spawn(async move {
         let layer = ServiceBuilder::new()
-            .timeout(Duration::from_secs(10))
-            .layer(TraceLayer::new_for_grpc());
+            .layer(BarrierLayer::new(barrier))
+            .layer(TraceLayer::new_for_grpc())
+            .into_inner();
 
         tonic::transport::Server::builder()
-            .trace_fn(|_| tracing::debug_span!("message_hub"))
             .layer(layer)
+            .trace_fn(|_| tracing::debug_span!("message_hub"))
+            .timeout(Duration::from_secs(10))
             .add_service(MessageHubServer::new(message_hub))
             .serve((Ipv6Addr::LOCALHOST, GRPC_PORT).into())
             .await
@@ -92,7 +100,7 @@ pub async fn init_message_hub(
 
     // Event loop for the astarte device sdk
     let event_loop = tasks.spawn(async move {
-        device
+        connection
             .handle_events()
             .await
             .wrap_err("disconnected from astarte")
@@ -111,4 +119,69 @@ pub async fn init_message_hub(
         event_loop,
         forwarder,
     })
+}
+
+#[derive(Debug, Clone)]
+struct BarrierLayer {
+    barrier: Arc<Barrier>,
+}
+
+impl BarrierLayer {
+    fn new(barrier: Arc<Barrier>) -> Self {
+        Self { barrier }
+    }
+}
+
+impl<S> Layer<S> for BarrierLayer {
+    type Service = BarrierService<S>;
+
+    fn layer(&self, inner: S) -> Self::Service {
+        BarrierService {
+            inner,
+            barrier: self.barrier.clone(),
+        }
+    }
+}
+
+/// Wait for the client to send a message and sync after the response is completed.
+#[derive(Debug, Clone)]
+struct BarrierService<S> {
+    barrier: Arc<Barrier>,
+    inner: S,
+}
+
+impl<S, R> Service<R> for BarrierService<S>
+where
+    S: Service<R> + Clone + Send + 'static,
+    S::Future: Send,
+    S::Response: Send,
+    S::Error: Send,
+    R: Send + 'static,
+{
+    type Response = S::Response;
+
+    type Error = S::Error;
+
+    type Future = BoxFuture<'static, Result<S::Response, S::Error>>;
+
+    fn poll_ready(
+        &mut self,
+        cx: &mut std::task::Context<'_>,
+    ) -> std::task::Poll<Result<(), Self::Error>> {
+        self.inner.poll_ready(cx)
+    }
+
+    fn call(&mut self, req: R) -> Self::Future {
+        let service = self.clone();
+        let mut service = std::mem::replace(self, service);
+
+        async move {
+            let res = service.inner.call(req).await;
+
+            service.barrier.wait().await;
+
+            res
+        }
+        .boxed()
+    }
 }