+++ CHANGELOG.md
@@ -6,19 +6,26 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+
 ### Changed
-- Update Astarte Device Sdk to 0.7.1 release.
+
+- Update Astarte Device Sdk to 0.8.1 release.
 - Bump MSRV to 1.72.0.
 
 ## [0.5.3] - 2024-01-31
+
 ### Added
+
 - Option to configure the timeout and keep alive interval for the MQTT connection to astarte.
 
 ## [0.5.2] - 2023-07-03
+
 ### Added
+
 - Add support to receive `device_id` option from dbus.
 
 ### Changed
+
 - Make `device_id` options as optional field.
 
 ## [0.5.1] - 2023-04-26